requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
pydub
pydantic
openai
python-dotenv
boto3
requests
spacy
python-multipart
google-api-python-client 
google-auth-oauthlib 
google-auth-httplib2
google-cloud-texttospeech
psutil